+++ tutorials/plot_object_annotations.py
@@ -2,10 +2,12 @@
 The **events** and :class:`Annotations <mne.Annotations>` data structures
 =========================================================================
 
-Events and annotations are quite similar. This tutorial highlights their
-differences and similitudes and tries to shade some light to which one is
-preferred to use in different situations when using MNE.
-Here follows both terms definition from the :ref:`glossary`.
+Events and :class:`Annotations <mne.Annotations>` are quite similar.
+This tutorial highlights their differences and similarities, and tries to shed
+some light on which one is preferred to use in different situations when using
+MNE.
+
+Here are the definitions from the :ref:`glossary`.
 
     events
         Events correspond to specific time points in raw data; e.g., triggers,
@@ -15,25 +17,32 @@ Here follows both terms definition from the :ref:`glossary`.
         channel.
 
     annotations
-        One annotation is defined by an onset, a duration and a string
+        An annotation is defined by an onset, a duration, and a string
         description. It can contain information about the experiments, but
         also details on signals marked by a human: bad data segments,
         sleep scores, sleep events (spindles, K-complex) etc.
 
-They both can be seen as triplets where the first element answers to **when**
-something happens and the last element refers to **what** is it. The
-main differnce is that events the when is samples with respect to the first
-sample and the what is an integer id; while in annotations the when is in
-seconds with respect to an origin and the what is an arbitrary string.
-The second element of the triplets have no direct relation between the two
-structures. For the events case, the second element corresponds to id of the
-previous active event. Whereas, the second element of the
-:class:`mne.Annotations` is a float indicating its duration in seconds.
+Both events and :class:`Annotations <mne.Annotations>` be seen as triplets
+where the first element answers to **when** something happens and the last
+element refers to **what** is it.
+The main difference is that events represent the onset in samples relative to
+the first sample value (:attr:`raw.first_samp <mne.io.Raw.first_samp>`), and
+the description is an integer value.
+In contrast, :class:`Annotations <mne.Annotations>` represents the
+``onset`` in seconds (relative to the reference ``orig_time``),
+ and the ``description`` is an arbitrary string.
+There is no correspondence between the second element of events and
+:class:`Annotations <mne.Annotations>`.
+For events, the second element corresponds to the ID of the previously active
+event.
+The second element of :class:`Annotations <mne.Annotations>` is a float
+indicating its duration in seconds.
 
 See :ref:`sphx_glr_auto_examples_io_plot_read_events.py`
-for a complete example in how to read, select and visualize **events**;
+for a complete example of how to read, select, and visualize **events**;
 and :ref:`sphx_glr_auto_tutorials_plot_artifacts_correction_rejection.py` to
-know how :class:`mne.Annotations` are used to mark bad segments of data.
+learn how :class:`Annotations <mne.Annotations>` are used to mark bad segments
+of data.
 
 An example of events and annotations
 ------------------------------------
@@ -44,15 +53,16 @@ marks bad segments due to eye blinks.
 
 import os.path as op
 import numpy as np
+from datetime import datetime
 
 import mne
 
-# load the data
+# Load the data
 data_path = mne.datasets.sample.data_path()
 fname = op.join(data_path, 'MEG', 'sample', 'sample_audvis_raw.fif')
 raw = mne.io.read_raw_fif(fname)
 
-# plot the events
+# Plot the events
 events = mne.find_events(raw)
 
 # Specify event_id dictionary based on the experiment
@@ -64,7 +74,7 @@ color = {1: 'green', 2: 'yellow', 3: 'red', 4: 'c', 5: 'black', 32: 'blue'}
 mne.viz.plot_events(events, raw.info['sfreq'], raw.first_samp, color=color,
                     event_id=event_id)
 
-# create some annotations
+# Create some annotations
 annotated_blink_raw = raw.copy()
 eog_events = mne.preprocessing.find_eog_events(raw)
 n_blinks = len(eog_events)
@@ -82,57 +92,78 @@ annotated_blink_raw.plot()  # plot the annotated raw
 # Working with Annotations
 # ------------------------
 #
-# An important element of the :class:`mne.Annotations` is ``orig_time`` which
-# is the time reference for the ``onset``. It is key to understand that when
-# calling `raw.set_annotation`, the given annotations is copied and transformed
-# so that `raw.annotations.orig_time` matches the recording time of the raw
-# object. (check :class:`mne.Annotations` documentation notes to see the
-# expected behavior depending of `meas_date` and `orig_time`. Notice that
-# `meas_date` is the :class:`Info <mne.Info>` attribute of the recording time.
-# Find more in :ref:`sphx_glr_auto_tutorials_plot_info.py`)
+# An important element of :class:`Annotations <mne.Annotations>` is
+# ``orig_time`` which is the time reference for the ``onset``.
+# It is key to understand that when calling
+# :func:`raw.set_annotations <mne.io.Raw.set_annotations>`, the given
+# annotations are copied and transformed so that
+# :class:`raw.annotations.orig_time <mne.Annotations>`
+# matches the recording time of the raw object.
+# Refer to the documentation of :class:`Annotations <mne.Annotations>` to see
+# the expected behavior depending on ``meas_date`` and ``orig_time``.
+# Where ``meas_date`` is the recording time of the stored in
+# :class:`Info <mne.Info>`.
+# You can find more information about :class:`Info <mne.Info>` in
+# :ref:`sphx_glr_auto_tutorials_plot_info.py`.
 #
-# We'll now manipulate some simulated annotations objects.
-#
-# First let's create an annotation object without orig_time. It this case
-# one assumes that the orig_time is the time of the first sample of data.
+# We'll now manipulate some simulated annotations.
+# The first annotations has ``orig_time`` set to ``None`` while the
+# second is set to a chosen POSIX timestamp for illustration purposes.
+
+###############################################################################
+
+# Create an annotation object without orig_time
 annot_none = mne.Annotations(onset=[0, 2, 9], duration=[0.5, 4, 0],
                              description=['foo', 'bar', 'foo'],
                              orig_time=None)
 print(annot_none)
 
+# Create an annotation object with orig_time
+orig_time = '2002-12-03 19:01:31.676071'
+my_datetime = datetime.strptime(orig_time, '%Y-%m-%d %H:%M:%S.%f')
+posix_timestamp = (my_datetime - datetime(1970, 1, 1)).total_seconds()
+print('{0:f} is the POSIX timestamp of {1:s}'.format(posix_timestamp,
+                                                     orig_time))
 
-###############################################################################
-
-# Now let's create annotation object with orig_time
 annot_orig = mne.Annotations(onset=[22, 24, 31], duration=[0.5, 4, 0],
                              description=['foo', 'bar', 'foo'],
                              orig_time=1038942091.6760709)
 print(annot_orig)
 
 ###############################################################################
+# Now we create two raw objects, set the annotations and plot them to compare
+# them.
 
-# create two cropped copies of raw with the two previous annotations
+# Create two cropped copies of raw with the two previous annotations
 raw_a = raw.copy().crop(tmax=12).set_annotations(annot_none)
 raw_b = raw.copy().crop(tmax=12).set_annotations(annot_orig)
 
-# plot the raw objects
+# Plot the raw objects
 raw_a.plot()
 raw_b.plot()
 
-###############################################################################
-
-# show the annotations in the raw objects
+# Show the annotations in the raw objects
 print(raw_a.annotations)
 print(raw_b.annotations)
 
-###############################################################################
-# show that the onsets are the same
+# Show that the onsets are the same
 print(raw_a.annotations.onset)
 print(raw_b.annotations.onset)
 
 ###############################################################################
 #
-# It is possible to concatenate two annotations with the + like for lists.
+# Notice that for the case where ``orig_time`` is ``None``,
+# one assumes that the orig_time is the time of the first sample of data.
+
+raw_delta = (1 / raw.info['sfreq'])
+print('raw.first_sample is {}'.format(raw.first_samp * raw_delta))
+print('annot_none.onset[0] is {}'.format(annot_none.onset[0]))
+print('raw_a.annotations.onset[0] is {}'.format(raw_a.annotations.onset[0]))
+
+###############################################################################
+#
+# It is possible to concatenate two annotations with the + operator like for
+# lists if both share the same ``orig_time``
 
 annot = mne.Annotations(onset=[10], duration=[0.5],
                         description=['foobar'],